# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kailali-4 Connect - MP Office Digital Governance Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ComplaintStatus =
  (typeof ComplaintStatus)[keyof typeof ComplaintStatus];

export const ComplaintStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface Complaint {
  id: number;
  name: string;
  phone?: string;
  palika: string;
  ward: number;
  category: string;
  description: string;
  status: ComplaintStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateComplaintBody {
  name: string;
  phone?: string;
  palika: string;
  ward: number;
  category: string;
  description: string;
}

export type UpdateComplaintStatusBodyStatus =
  (typeof UpdateComplaintStatusBodyStatus)[keyof typeof UpdateComplaintStatusBodyStatus];

export const UpdateComplaintStatusBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface UpdateComplaintStatusBody {
  status: UpdateComplaintStatusBodyStatus;
}

export interface Idea {
  id: number;
  title: string;
  description: string;
  submittedBy: string;
  ward: number;
  upvotes: number;
  createdAt: string;
}

export interface CreateIdeaBody {
  title: string;
  description: string;
  submittedBy: string;
  ward: number;
}

export interface NewsItem {
  id: number;
  title: string;
  content: string;
  category: string;
  publishedAt: string;
}

export interface CreateNewsBody {
  title: string;
  content: string;
  category: string;
}

export interface DashboardStats {
  total: number;
  pending: number;
  inProgress: number;
  resolved: number;
  totalIdeas: number;
  totalNews: number;
}

export interface WardStat {
  ward: number;
  total: number;
  resolved: number;
}

export interface MpProfile {
  name: string;
  message: string;
  photoUrl?: string;
}

export interface DeleteAllResult {
  deleted: number;
}

export type TeamMemberRank =
  (typeof TeamMemberRank)[keyof typeof TeamMemberRank];

export const TeamMemberRank = {
  volunteer: "volunteer",
  coordinator: "coordinator",
  leader: "leader",
} as const;

export interface TeamMember {
  id: number;
  cid: string;
  name: string;
  phone?: string;
  palika: string;
  ward: number;
  photoUrl?: string;
  rank: TeamMemberRank;
  createdAt: string;
}

export interface JoinTeamBody {
  name: string;
  phone?: string;
  palika: string;
  ward: number;
  photoUrl?: string;
}

export type UpdateRankBodyRank =
  (typeof UpdateRankBodyRank)[keyof typeof UpdateRankBodyRank];

export const UpdateRankBodyRank = {
  volunteer: "volunteer",
  coordinator: "coordinator",
  leader: "leader",
} as const;

export interface UpdateRankBody {
  rank: UpdateRankBodyRank;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  complaint: "complaint",
  idea: "idea",
  news: "news",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  timestamp: string;
}

export type ListComplaintsParams = {
  status?: ListComplaintsStatus;
  ward?: number;
};

export type ListComplaintsStatus =
  (typeof ListComplaintsStatus)[keyof typeof ListComplaintsStatus];

export const ListComplaintsStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export type ListTeamMembersParams = {
  palika?: string;
  rank?: ListTeamMembersRank;
};

export type ListTeamMembersRank =
  (typeof ListTeamMembersRank)[keyof typeof ListTeamMembersRank];

export const ListTeamMembersRank = {
  volunteer: "volunteer",
  coordinator: "coordinator",
  leader: "leader",
} as const;