# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kailali-4 Connect - MP Office Digital Governance Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: complaints
    description: Citizen complaints
  - name: ideas
    description: Community ideas
  - name: news
    description: Official news and updates
  - name: dashboard
    description: Dashboard statistics and summaries
  - name: team
    description: Team member registry
  - name: settings
    description: Site settings and configuration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List all complaints
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, resolved]
        - name: ward
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintBody"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get a complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    patch:
      operationId: updateComplaintStatus
      tags: [complaints]
      summary: Update complaint status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintStatusBody"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  /ideas:
    get:
      operationId: listIdeas
      tags: [ideas]
      summary: List community ideas
      responses:
        "200":
          description: List of ideas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
    post:
      operationId: createIdea
      tags: [ideas]
      summary: Submit a new idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdeaBody"
      responses:
        "201":
          description: Idea created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"

  /ideas/{id}/upvote:
    post:
      operationId: upvoteIdea
      tags: [ideas]
      summary: Upvote an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated idea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"

  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List official news updates
      responses:
        "200":
          description: List of news items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsBody"
      responses:
        "201":
          description: News item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get complaint statistics summary
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/ward-breakdown:
    get:
      operationId: getWardBreakdown
      tags: [dashboard]
      summary: Get complaint counts grouped by ward
      responses:
        "200":
          description: Ward breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WardStat"

  /settings/mp-profile:
    get:
      operationId: getMpProfile
      tags: [settings]
      summary: Get MP profile settings
      responses:
        "200":
          description: MP profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpProfile"
    patch:
      operationId: updateMpProfile
      tags: [settings]
      summary: Update MP profile settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpProfile"
      responses:
        "200":
          description: Updated MP profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpProfile"

  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List all team members
      parameters:
        - name: palika
          in: query
          schema:
            type: string
        - name: rank
          in: query
          schema:
            type: string
            enum: [volunteer, coordinator, leader]
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: joinTeam
      tags: [team]
      summary: Register a new team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinTeamBody"
      responses:
        "201":
          description: Team member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"

  /team/all:
    delete:
      operationId: deleteAllTeamMembers
      tags: [team]
      summary: Delete all team members
      responses:
        "200":
          description: All team members deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAllResult"

  /complaints/all:
    delete:
      operationId: deleteAllComplaints
      tags: [complaints]
      summary: Delete all complaints
      responses:
        "200":
          description: All complaints deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAllResult"

  /team/{id}/rank:
    patch:
      operationId: updateTeamRank
      tags: [team]
      summary: Update a team member's rank
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRankBody"
      responses:
        "200":
          description: Rank updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Member not found

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Complaint:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        palika:
          type: string
        ward:
          type: integer
        category:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, resolved]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - palika
        - ward
        - category
        - description
        - status
        - createdAt
        - updatedAt

    CreateComplaintBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        palika:
          type: string
        ward:
          type: integer
        category:
          type: string
        description:
          type: string
      required:
        - name
        - palika
        - ward
        - category
        - description

    UpdateComplaintStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, resolved]
      required:
        - status

    Idea:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        submittedBy:
          type: string
        ward:
          type: integer
        upvotes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - submittedBy
        - ward
        - upvotes
        - createdAt

    CreateIdeaBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        submittedBy:
          type: string
        ward:
          type: integer
      required:
        - title
        - description
        - submittedBy
        - ward

    NewsItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        publishedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - category
        - publishedAt

    CreateNewsBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
      required:
        - title
        - content
        - category

    DashboardStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        resolved:
          type: integer
        totalIdeas:
          type: integer
        totalNews:
          type: integer
      required:
        - total
        - pending
        - inProgress
        - resolved
        - totalIdeas
        - totalNews

    WardStat:
      type: object
      properties:
        ward:
          type: integer
        total:
          type: integer
        resolved:
          type: integer
      required:
        - ward
        - total
        - resolved

    MpProfile:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
        photoUrl:
          type: string
      required:
        - name
        - message

    DeleteAllResult:
      type: object
      properties:
        deleted:
          type: integer
      required:
        - deleted

    TeamMember:
      type: object
      properties:
        id:
          type: integer
        cid:
          type: string
        name:
          type: string
        phone:
          type: string
        palika:
          type: string
        ward:
          type: integer
        photoUrl:
          type: string
        rank:
          type: string
          enum: [volunteer, coordinator, leader]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - cid
        - name
        - palika
        - ward
        - rank
        - createdAt

    JoinTeamBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        palika:
          type: string
        ward:
          type: integer
        photoUrl:
          type: string
      required:
        - name
        - palika
        - ward

    UpdateRankBody:
      type: object
      properties:
        rank:
          type: string
          enum: [volunteer, coordinator, leader]
      required:
        - rank

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [complaint, idea, news]
        title:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - description
        - timestamp