# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kailali-4 Connect - MP Office Digital Governance Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all complaints
 */
export const ListComplaintsQueryParams = zod.object({
  status: zod.enum(["pending", "in_progress", "resolved"]).optional(),
  ward: zod.coerce.number().optional(),
});

export const ListComplaintsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string().optional(),
  palika: zod.string(),
  ward: zod.number(),
  category: zod.string(),
  description: zod.string(),
  status: zod.enum(["pending", "in_progress", "resolved"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListComplaintsResponse = zod.array(ListComplaintsResponseItem);

/**
 * @summary Submit a new complaint
 */
export const CreateComplaintBody = zod.object({
  name: zod.string(),
  phone: zod.string().optional(),
  palika: zod.string(),
  ward: zod.number(),
  category: zod.string(),
  description: zod.string(),
});

/**
 * @summary Get a complaint by ID
 */
export const GetComplaintParams = zod.object({
  id: zod.coerce.number(),
});

export const GetComplaintResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string().optional(),
  palika: zod.string(),
  ward: zod.number(),
  category: zod.string(),
  description: zod.string(),
  status: zod.enum(["pending", "in_progress", "resolved"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update complaint status
 */
export const UpdateComplaintStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateComplaintStatusBody = zod.object({
  status: zod.enum(["pending", "in_progress", "resolved"]),
});

export const UpdateComplaintStatusResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string().optional(),
  palika: zod.string(),
  ward: zod.number(),
  category: zod.string(),
  description: zod.string(),
  status: zod.enum(["pending", "in_progress", "resolved"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List community ideas
 */
export const ListIdeasResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  submittedBy: zod.string(),
  ward: zod.number(),
  upvotes: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListIdeasResponse = zod.array(ListIdeasResponseItem);

/**
 * @summary Submit a new idea
 */
export const CreateIdeaBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  submittedBy: zod.string(),
  ward: zod.number(),
});

/**
 * @summary Upvote an idea
 */
export const UpvoteIdeaParams = zod.object({
  id: zod.coerce.number(),
});

export const UpvoteIdeaResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  submittedBy: zod.string(),
  ward: zod.number(),
  upvotes: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List official news updates
 */
export const ListNewsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  category: zod.string(),
  publishedAt: zod.coerce.date(),
});
export const ListNewsResponse = zod.array(ListNewsResponseItem);

/**
 * @summary Create a news update
 */
export const CreateNewsBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  category: zod.string(),
});

/**
 * @summary Get complaint statistics summary
 */
export const GetDashboardStatsResponse = zod.object({
  total: zod.number(),
  pending: zod.number(),
  inProgress: zod.number(),
  resolved: zod.number(),
  totalIdeas: zod.number(),
  totalNews: zod.number(),
});

/**
 * @summary Get complaint counts grouped by ward
 */
export const GetWardBreakdownResponseItem = zod.object({
  ward: zod.number(),
  total: zod.number(),
  resolved: zod.number(),
});
export const GetWardBreakdownResponse = zod.array(GetWardBreakdownResponseItem);

/**
 * @summary Get MP profile settings
 */
export const GetMpProfileResponse = zod.object({
  name: zod.string(),
  message: zod.string(),
  photoUrl: zod.string().optional(),
});

/**
 * @summary Update MP profile settings
 */
export const UpdateMpProfileBody = zod.object({
  name: zod.string(),
  message: zod.string(),
  photoUrl: zod.string().optional(),
});

export const UpdateMpProfileResponse = zod.object({
  name: zod.string(),
  message: zod.string(),
  photoUrl: zod.string().optional(),
});

/**
 * @summary List all team members
 */
export const ListTeamMembersQueryParams = zod.object({
  palika: zod.coerce.string().optional(),
  rank: zod.enum(["volunteer", "coordinator", "leader"]).optional(),
});

export const ListTeamMembersResponseItem = zod.object({
  id: zod.number(),
  cid: zod.string(),
  name: zod.string(),
  phone: zod.string().optional(),
  palika: zod.string(),
  ward: zod.number(),
  photoUrl: zod.string().optional(),
  rank: zod.enum(["volunteer", "coordinator", "leader"]),
  createdAt: zod.coerce.date(),
});
export const ListTeamMembersResponse = zod.array(ListTeamMembersResponseItem);

/**
 * @summary Register a new team member
 */
export const JoinTeamBody = zod.object({
  name: zod.string(),
  phone: zod.string().optional(),
  palika: zod.string(),
  ward: zod.number(),
  photoUrl: zod.string().optional(),
});

/**
 * @summary Delete all team members
 */
export const DeleteAllTeamMembersResponse = zod.object({
  deleted: zod.number(),
});

/**
 * @summary Delete all complaints
 */
export const DeleteAllComplaintsResponse = zod.object({
  deleted: zod.number(),
});

/**
 * @summary Update a team member's rank
 */
export const UpdateTeamRankParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTeamRankBody = zod.object({
  rank: zod.enum(["volunteer", "coordinator", "leader"]),
});

export const UpdateTeamRankResponse = zod.object({
  id: zod.number(),
  cid: zod.string(),
  name: zod.string(),
  phone: zod.string().optional(),
  palika: zod.string(),
  ward: zod.number(),
  photoUrl: zod.string().optional(),
  rank: zod.enum(["volunteer", "coordinator", "leader"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get recent activity feed
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["complaint", "idea", "news"]),
  title: zod.string(),
  description: zod.string(),
  timestamp: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);