# lib/api-zod/src/generated/types/mpProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kailali-4 Connect - MP Office Digital Governance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MpProfile {
  name: string;
  message: string;
  photoUrl?: string;
}